Make Travis-CI PHP example to support GitLab-CI shared runners

+++ Tests/HelloWorldTest.php
@@ -16,6 +16,8 @@ class HelloWorldTest extends PHPUnit_Framework_TestCase
 
     public function tearDown()
     {
+        if (!$this->pdo)
+            return;
         $this->pdo->query("DROP TABLE hello");
     }
